fix: move getTsRanges to utils
